requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pymysql==1.1.1
python-dotenv==1.1.0
